refactor(runtime): extract issue state store module (#1284)
